LayoutTests/fast/forms/ValidityState-stepMismatch-unsupported-expected.txt

Check stepMismatch results for types which does not support for step.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS input.validity.stepMismatch is false
PASS successfullyParsed is true

TEST COMPLETE
